SYS-2832 : UI updates (#2)
WHAT
- Implements the design changes suggested
- Implements account disconnect functionality without leaving the dapp
- Implemented alerts prior to every signature request.
- Added lower support for ERC777 and ETH tokens
- Added balance query support for AVT and ETH tokens
- The form on the Query & Withdraw page accepts the sender's aventus
address, public key or the recipient's ethereum address.
- Add accordion and pagination to handle lower data pulled from the
backend.
- Added text support for Nova wallet.

WHY
- Although minor changes, might offer a better user experience. 
- Introduces the ability to disconnect an account from the dapp without
having to do it from the extension.
- To inform the user why the signature is required. 
- The lifting dapp lifts ERC20, ERC777 and ETH tokens, so the lowering
dapp should also allow for lowering them.
- Allows the user to query the balance for other token types that exist
on the AVN
- Ensures whatever the user chooses to enter in, irregardless of the
ss58 format, the input can be used, if valid
- The accordion ensures the user is not overloaded with data and can
view one lower at a time. The pagination ensures the page is not too
long if flooded with multiple lower requests plus the user has a limited
number of lowers per page to filter through.
- The dapp is already compatible with the Nova wallet. The text informs
the user on which button works with the wallet.